Task3/src/services/service-b/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
opentelemetry-api>=1.28.0,<2
opentelemetry-sdk>=1.28.0,<2
opentelemetry-exporter-otlp-proto-http>=1.28.0,<2
opentelemetry-instrumentation-fastapi>=0.49b0,<1
